+++ packages/rest-api/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.68](https://github.com/synapsecns/sanguine/compare/@synapsecns/rest-api@1.0.67...@synapsecns/rest-api@1.0.68) (2024-07-24)
+
+**Note:** Version bump only for package @synapsecns/rest-api
+
+
+
+
+
+## [1.0.67](https://github.com/synapsecns/sanguine/compare/@synapsecns/rest-api@1.0.66...@synapsecns/rest-api@1.0.67) (2024-07-15)
+
+**Note:** Version bump only for package @synapsecns/rest-api
+
+
+
+
+
 ## [1.0.66](https://github.com/synapsecns/sanguine/compare/@synapsecns/rest-api@1.0.65...@synapsecns/rest-api@1.0.66) (2024-07-03)
 
 **Note:** Version bump only for package @synapsecns/rest-api